port TPCH: proof of concept using dialects and procedures
