+++ src/com/oltpbenchmark/benchmarks/tpch/procedures/Q11.java
@@ -14,35 +14,40 @@
  *  limitations under the License.                                            *
  ******************************************************************************/
 
-package com.oltpbenchmark.benchmarks.tpch.queries;
+package com.oltpbenchmark.benchmarks.tpch.procedures;
 
 import com.oltpbenchmark.api.SQLStmt;
 
-public class Q3 extends GenericQuery {
+public class Q11 extends GenericQuery {
 
     public final SQLStmt query_stmt = new SQLStmt(
               "select "
-            +     "l_orderkey, "
-            +     "sum(l_extendedprice * (1 - l_discount)) as revenue, "
-            +     "o_orderdate, "
-            +     "o_shippriority "
+            +     "ps_partkey, "
+            +     "sum(ps_supplycost * ps_availqty) as value "
             + "from "
-            +     "customer, "
-            +     "orders, "
-            +     "lineitem "
+            +     "partsupp, "
+            +     "supplier, "
+            +     "nation "
             + "where "
-            +     "c_mktsegment = 'MACHINERY' "
-            +     "and c_custkey = o_custkey "
-            +     "and l_orderkey = o_orderkey "
-            +     "and o_orderdate < date '1995-03-10' "
-            +     "and l_shipdate > date '1995-03-10' "
+            +     "ps_suppkey = s_suppkey "
+            +     "and s_nationkey = n_nationkey "
+            +     "and n_name = 'ETHIOPIA' "
             + "group by "
-            +     "l_orderkey, "
-            +     "o_orderdate, "
-            +     "o_shippriority "
+            +     "ps_partkey having "
+            +         "sum(ps_supplycost * ps_availqty) > ( "
+            +             "select "
+            +                 "sum(ps_supplycost * ps_availqty) * 0.0000003333 "
+            +             "from "
+            +                 "partsupp, "
+            +                 "supplier, "
+            +                 "nation "
+            +             "where "
+            +                 "ps_suppkey = s_suppkey "
+            +                 "and s_nationkey = n_nationkey "
+            +                 "and n_name = 'ETHIOPIA' "
+            +         ") "
             + "order by "
-            +     "revenue desc, "
-            +     "o_orderdate"
+            +     "value desc"
         );
 
     protected SQLStmt get_query() {
