+++ src/com/oltpbenchmark/benchmarks/tpch/procedures/Q2.java
@@ -14,50 +14,56 @@
  *  limitations under the License.                                            *
  ******************************************************************************/
 
-package com.oltpbenchmark.benchmarks.tpch.queries;
+package com.oltpbenchmark.benchmarks.tpch.procedures;
 
 import com.oltpbenchmark.api.SQLStmt;
 
-public class Q8 extends GenericQuery {
+public class Q2 extends GenericQuery {
 
     public final SQLStmt query_stmt = new SQLStmt(
               "select "
-            +     "o_year, "
-            +     "sum(case "
-            +         "when nation = 'CANADA' then volume "
-            +         "else 0 "
-            +     "end) / sum(volume) as mkt_share "
+            +     "s_acctbal, "
+            +     "s_name, "
+            +     "n_name, "
+            +     "p_partkey, "
+            +     "p_mfgr, "
+            +     "s_address, "
+            +     "s_phone, "
+            +     "s_comment "
             + "from "
-            +     "( "
+            +     "part, "
+            +     "supplier, "
+            +     "partsupp, "
+            +     "nation, "
+            +     "region "
+            + "where "
+            +     "p_partkey = ps_partkey "
+            +     "and s_suppkey = ps_suppkey "
+            +     "and p_size = 42 "
+            +     "and p_type like '%STEEL' "
+            +     "and s_nationkey = n_nationkey "
+            +     "and n_regionkey = r_regionkey "
+            +     "and r_name = 'AMERICA' "
+            +     "and ps_supplycost = ( "
             +         "select "
-            +             "extract(year from o_orderdate) as o_year, "
-            +             "l_extendedprice * (1 - l_discount) as volume, "
-            +             "n2.n_name as nation "
+            +             "min(ps_supplycost) "
             +         "from "
-            +             "part, "
+            +             "partsupp, "
             +             "supplier, "
-            +             "lineitem, "
-            +             "orders, "
-            +             "customer, "
-            +             "nation n1, "
-            +             "nation n2, "
+            +             "nation, "
             +             "region "
             +         "where "
-            +             "p_partkey = l_partkey "
-            +             "and s_suppkey = l_suppkey "
-            +             "and l_orderkey = o_orderkey "
-            +             "and o_custkey = c_custkey "
-            +             "and c_nationkey = n1.n_nationkey "
-            +             "and n1.n_regionkey = r_regionkey "
+            +             "p_partkey = ps_partkey "
+            +             "and s_suppkey = ps_suppkey "
+            +             "and s_nationkey = n_nationkey "
+            +             "and n_regionkey = r_regionkey "
             +             "and r_name = 'AMERICA' "
-            +             "and s_nationkey = n2.n_nationkey "
-            +             "and o_orderdate between date '1995-01-01' and date '1996-12-31' "
-            +             "and p_type = 'ECONOMY POLISHED STEEL' "
-            +     ") as all_nations "
-            + "group by "
-            +     "o_year "
+            +     ") "
             + "order by "
-            +     "o_year"
+            +     "s_acctbal desc, "
+            +     "n_name, "
+            +     "s_name, "
+            +     "p_partkey"
         );
 
     protected SQLStmt get_query() {
