+++ src/com/oltpbenchmark/benchmarks/tpch/procedures/Q21.java
@@ -14,56 +14,52 @@
  *  limitations under the License.                                            *
  ******************************************************************************/
 
-package com.oltpbenchmark.benchmarks.tpch.queries;
+package com.oltpbenchmark.benchmarks.tpch.procedures;
 
 import com.oltpbenchmark.api.SQLStmt;
 
-public class Q2 extends GenericQuery {
+public class Q21 extends GenericQuery {
 
     public final SQLStmt query_stmt = new SQLStmt(
               "select "
-            +     "s_acctbal, "
             +     "s_name, "
-            +     "n_name, "
-            +     "p_partkey, "
-            +     "p_mfgr, "
-            +     "s_address, "
-            +     "s_phone, "
-            +     "s_comment "
+            +     "count(*) as numwait "
             + "from "
-            +     "part, "
             +     "supplier, "
-            +     "partsupp, "
-            +     "nation, "
-            +     "region "
+            +     "lineitem l1, "
+            +     "orders, "
+            +     "nation "
             + "where "
-            +     "p_partkey = ps_partkey "
-            +     "and s_suppkey = ps_suppkey "
-            +     "and p_size = 42 "
-            +     "and p_type like '%STEEL' "
-            +     "and s_nationkey = n_nationkey "
-            +     "and n_regionkey = r_regionkey "
-            +     "and r_name = 'AMERICA' "
-            +     "and ps_supplycost = ( "
+            +     "s_suppkey = l1.l_suppkey "
+            +     "and o_orderkey = l1.l_orderkey "
+            +     "and o_orderstatus = 'F' "
+            +     "and l1.l_receiptdate > l1.l_commitdate "
+            +     "and exists ( "
             +         "select "
-            +             "min(ps_supplycost) "
+            +             "* "
             +         "from "
-            +             "partsupp, "
-            +             "supplier, "
-            +             "nation, "
-            +             "region "
+            +             "lineitem l2 "
             +         "where "
-            +             "p_partkey = ps_partkey "
-            +             "and s_suppkey = ps_suppkey "
-            +             "and s_nationkey = n_nationkey "
-            +             "and n_regionkey = r_regionkey "
-            +             "and r_name = 'AMERICA' "
+            +             "l2.l_orderkey = l1.l_orderkey "
+            +             "and l2.l_suppkey <> l1.l_suppkey "
             +     ") "
+            +     "and not exists ( "
+            +         "select "
+            +             "* "
+            +         "from "
+            +             "lineitem l3 "
+            +         "where "
+            +             "l3.l_orderkey = l1.l_orderkey "
+            +             "and l3.l_suppkey <> l1.l_suppkey "
+            +             "and l3.l_receiptdate > l3.l_commitdate "
+            +     ") "
+            +     "and s_nationkey = n_nationkey "
+            +     "and n_name = 'SAUDI ARABIA' "
+            + "group by "
+            +     "s_name "
             + "order by "
-            +     "s_acctbal desc, "
-            +     "n_name, "
-            +     "s_name, "
-            +     "p_partkey"
+            +     "numwait desc, "
+            +     "s_name"
         );
 
     protected SQLStmt get_query() {
