+++ src/com/oltpbenchmark/benchmarks/tpch/procedures/Q22.java
@@ -14,52 +14,50 @@
  *  limitations under the License.                                            *
  ******************************************************************************/
 
-package com.oltpbenchmark.benchmarks.tpch.queries;
+package com.oltpbenchmark.benchmarks.tpch.procedures;
 
 import com.oltpbenchmark.api.SQLStmt;
 
-public class Q21 extends GenericQuery {
+public class Q22 extends GenericQuery {
 
     public final SQLStmt query_stmt = new SQLStmt(
               "select "
-            +     "s_name, "
-            +     "count(*) as numwait "
+            +     "cntrycode, "
+            +     "count(*) as numcust, "
+            +     "sum(c_acctbal) as totacctbal "
             + "from "
-            +     "supplier, "
-            +     "lineitem l1, "
-            +     "orders, "
-            +     "nation "
-            + "where "
-            +     "s_suppkey = l1.l_suppkey "
-            +     "and o_orderkey = l1.l_orderkey "
-            +     "and o_orderstatus = 'F' "
-            +     "and l1.l_receiptdate > l1.l_commitdate "
-            +     "and exists ( "
+            +     "( "
             +         "select "
-            +             "* "
+            +             "substring(c_phone from 1 for 2) as cntrycode, "
+            +             "c_acctbal "
             +         "from "
-            +             "lineitem l2 "
+            +             "customer "
             +         "where "
-            +             "l2.l_orderkey = l1.l_orderkey "
-            +             "and l2.l_suppkey <> l1.l_suppkey "
-            +     ") "
-            +     "and not exists ( "
-            +         "select "
-            +             "* "
-            +         "from "
-            +             "lineitem l3 "
-            +         "where "
-            +             "l3.l_orderkey = l1.l_orderkey "
-            +             "and l3.l_suppkey <> l1.l_suppkey "
-            +             "and l3.l_receiptdate > l3.l_commitdate "
-            +     ") "
-            +     "and s_nationkey = n_nationkey "
-            +     "and n_name = 'SAUDI ARABIA' "
+            +             "substring(c_phone from 1 for 2) in "
+            +                 "('20', '32', '44', '33', '29', '22', '31') "
+            +             "and c_acctbal > ( "
+            +                 "select "
+            +                     "avg(c_acctbal) "
+            +                 "from "
+            +                     "customer "
+            +                 "where "
+            +                     "c_acctbal > 0.00 "
+            +                     "and substring(c_phone from 1 for 2) in "
+            +                         "('20', '32', '44', '33', '29', '22', '31') "
+            +             ") "
+            +             "and not exists ( "
+            +                 "select "
+            +                     "* "
+            +                 "from "
+            +                     "orders "
+            +                 "where "
+            +                     "o_custkey = c_custkey "
+            +             ") "
+            +     ") as custsale "
             + "group by "
-            +     "s_name "
+            +     "cntrycode "
             + "order by "
-            +     "numwait desc, "
-            +     "s_name"
+            +     "cntrycode"
         );
 
     protected SQLStmt get_query() {
