+++ src/com/oltpbenchmark/benchmarks/tpch/procedures/Q4.java
@@ -14,41 +14,34 @@
  *  limitations under the License.                                            *
  ******************************************************************************/
 
-package com.oltpbenchmark.benchmarks.tpch.queries;
+package com.oltpbenchmark.benchmarks.tpch.procedures;
 
 import com.oltpbenchmark.api.SQLStmt;
 
-public class Q12 extends GenericQuery {
+public class Q4 extends GenericQuery {
 
     public final SQLStmt query_stmt = new SQLStmt(
               "select "
-            +     "l_shipmode, "
-            +     "sum(case "
-            +         "when o_orderpriority = '1-URGENT' "
-            +             "or o_orderpriority = '2-HIGH' "
-            +             "then 1 "
-            +         "else 0 "
-            +     "end) as high_line_count, "
-            +     "sum(case "
-            +         "when o_orderpriority <> '1-URGENT' "
-            +             "and o_orderpriority <> '2-HIGH' "
-            +             "then 1 "
-            +         "else 0 "
-            +     "end) as low_line_count "
+            +     "o_orderpriority, "
+            +     "count(*) as order_count "
             + "from "
-            +     "orders, "
-            +     "lineitem "
+            +     "orders "
             + "where "
-            +     "o_orderkey = l_orderkey "
-            +     "and l_shipmode in ('AIR', 'REG AIR') "
-            +     "and l_commitdate < l_receiptdate "
-            +     "and l_shipdate < l_commitdate "
-            +     "and l_receiptdate >= date '1997-01-01' "
-            +     "and l_receiptdate < date '1997-01-01' + interval '1' year "
+            +     "o_orderdate >= date '1994-08-01' "
+            +     "and o_orderdate < date '1994-08-01' + interval '3' month "
+            +     "and exists ( "
+            +         "select "
+            +             "* "
+            +         "from "
+            +             "lineitem "
+            +         "where "
+            +             "l_orderkey = o_orderkey "
+            +             "and l_commitdate < l_receiptdate "
+            +     ") "
             + "group by "
-            +     "l_shipmode "
+            +     "o_orderpriority "
             + "order by "
-            +     "l_shipmode"
+            +     "o_orderpriority"
         );
 
     protected SQLStmt get_query() {
