+++ src/com/oltpbenchmark/benchmarks/tpch/procedures/Q6.java
@@ -14,44 +14,22 @@
  *  limitations under the License.                                            *
  ******************************************************************************/
 
-package com.oltpbenchmark.benchmarks.tpch.queries;
+package com.oltpbenchmark.benchmarks.tpch.procedures;
 
 import com.oltpbenchmark.api.SQLStmt;
 
-public class Q10 extends GenericQuery {
+public class Q6 extends GenericQuery {
 
     public final SQLStmt query_stmt = new SQLStmt(
               "select "
-            +     "c_custkey, "
-            +     "c_name, "
-            +     "sum(l_extendedprice * (1 - l_discount)) as revenue, "
-            +     "c_acctbal, "
-            +     "n_name, "
-            +     "c_address, "
-            +     "c_phone, "
-            +     "c_comment "
+            +     "sum(l_extendedprice * l_discount) as revenue "
             + "from "
-            +     "customer, "
-            +     "orders, "
-            +     "lineitem, "
-            +     "nation "
+            +     "lineitem "
             + "where "
-            +     "c_custkey = o_custkey "
-            +     "and l_orderkey = o_orderkey "
-            +     "and o_orderdate >= date '1994-12-01' "
-            +     "and o_orderdate < date '1994-12-01' + interval '3' month "
-            +     "and l_returnflag = 'R' "
-            +     "and c_nationkey = n_nationkey "
-            + "group by "
-            +     "c_custkey, "
-            +     "c_name, "
-            +     "c_acctbal, "
-            +     "c_phone, "
-            +     "n_name, "
-            +     "c_address, "
-            +     "c_comment "
-            + "order by "
-            +     "revenue desc"
+            +     "l_shipdate >= date '1997-01-01' "
+            +     "and l_shipdate < date '1997-01-01' + interval '1' year "
+            +     "and l_discount between 0.07 - 0.01 and 0.07 + 0.01 "
+            +     "and l_quantity < 24"
         );
 
     protected SQLStmt get_query() {
