+++ src/com/oltpbenchmark/benchmarks/tpch/procedures/query15-create-view.sql
