+++ src/com/oltpbenchmark/benchmarks/tpch/procedures/query15-drop-view.sql
