+++ src/com/oltpbenchmark/benchmarks/tpch/procedures/query19.sql
