+++ src/com/oltpbenchmark/benchmarks/tpch/procedures/query6.sql
